// src/openapi/model/updateCodeOfAutolivCommand.ts
/**
 * ICNC.ERP.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UpdateCodeOfAutolivCommand { 
    product_id?: string;
    file_name?: string | null;
    program_code?: string | null;
    machine_code?: string | null;
    workflow_id?: number | null;
}